Refactor - Put all Trello API Calls into Trello_API_Helper.py
