+++ todo_app/data/Trello.py
@@ -1,62 +1,21 @@
-import os
 import requests
 from dotenv import load_dotenv
+
+from todo_app.data.Trello_API_helper import get_member_details
+from todo_app.data.Trello_API_helper import get_board_details
+from todo_app.data.Trello_API_helper import get_list_for_card
+from todo_app.data.Trello_API_helper import get_lists_on_board
+from todo_app.data.Trello_API_helper import get_cards_on_list
+from todo_app.data.Trello_API_helper import get_cards_on_board
+from todo_app.data.Trello_API_helper import write_new_task_status
+from todo_app.data.Trello_API_helper import add_trello_task
+from todo_app.data.Trello_API_helper import delete_trello_task
+from todo_app.data.Trello_API_helper import run_get_query
+
 from todo_app.data.Trello_Member import Trello_Member
 from todo_app.data.Trello_Board import Trello_Board
 from todo_app.data.Trello_List import Trello_List
 from todo_app.data.Trello_Item import Trello_Item
-
-
-load_dotenv()
-TRELLO_KEY = os.getenv('TRELLO_KEY')
-TRELLO_TOKEN = os.getenv('TRELLO_TOKEN')
-
-headers = {
-   "Accept": "application/json"
-}
-
-query = {
-   'key' :  TRELLO_KEY,
-   'token' : TRELLO_TOKEN
-}
-
-def run_get_query(url):
-    return requests.request(
-        "GET",
-        url,
-        headers=headers,
-        params=query
-    )
-
-
-def get_member_details(user_id):
-    url = "https://api.trello.com/1/members/" + user_id
-    return run_get_query (url)
-
-def get_board_details(board_id):
-    url = "https://api.trello.com/1/boards/" + board_id
-    return run_get_query (url)
-
-def get_board_lists(board_id):
-    url = "ttps://api.trello.com/1/boards/" + board_id + "/lists"
-    return run_get_query (url)
-
-def get_cards_on_list(list_id):
-    url = "https://api.trello.com/1/lists/" + list_id + "/cards"
-    return run_get_query(url)
-
-def get_lists_on_board(board_id):
-    url = "https://api.trello.com/1/boards/" + board_id + "/lists"
-    return run_get_query(url)
-
-def get_cards_on_board(board_id):
-    url = "https://api.trello.com/1/boards/" + board_id + "/cards"
-    return run_get_query(url)
-
-def get_list_for_card(card_id):
-    url = "https://api.trello.com/1/cards/" + card_id + "/list"
-    response = run_get_query(url)
-    return response.json()['name']
 
 def get_member(member_id):
     response = get_member_details(member_id)
@@ -68,11 +27,9 @@ def get_card(task_id):
     card = response.json()
     return Trello_Item(card['id'], card['idShort'] , get_list_for_card(card['id']), card['name'], card['due'], card['idBoard'])
 
-
 def add_board_to_member(member,board_id):
     response = get_board_details(board_id)
     member.add_board(Trello_Board(response.json()['id'], response.json()['name']))
-    
     
 def add_lists_to_board(board):
     response = get_lists_on_board(board.id)
@@ -121,20 +78,6 @@ def get_list_by_list_name(board_id, name):
             return Trello_List(l['id'], l['name'])
     return None
 
-def write_new_task_status(task, list):
-    url = "https://api.trello.com/1/cards/" + task.id_long 
-    query = {
-        'key' :  TRELLO_KEY,
-        'token' : TRELLO_TOKEN,
-        'idList' : list.id
-    } 
-    response = requests.request(
-        "PUT",
-        url,
-        headers=headers,
-        params=query
-    )
-
 
 def update_task_status(id_long, action):
     task = get_card(id_long)
@@ -142,49 +85,17 @@ def update_task_status(id_long, action):
     new_list = get_list_by_list_name(task.board_id, new_list_name)
     write_new_task_status(task,new_list)
     
-
-    
-def delete_task_from_board(id_long):
-    url = "https://api.trello.com/1/cards/" + id_long 
-     
-    response = requests.request(
-        "DELETE",
-        url,
-        headers=headers,
-        params=query
-    )
-    print(response.text)
-
-
-def add_trello_task(title, target_date, board_id, list_id):
-    url = "https://api.trello.com/1/cards" 
-   
-    query = {
-      'key' :  TRELLO_KEY,
-      'token' : TRELLO_TOKEN,
-      'idList' : list_id,
-      'name': title,
-      'due': target_date,
-    } 
-    
-    response = requests.request(
-        "POST",
-        url,
-        headers=headers,
-        params=query
-    )
-    print(response.text)
-
 def add_task(title, target_date):
     member = get_member('alanjknight@hotmail.com')
     #assuming one board, and assume all new tickets go on todo list
-    
     get_boards_for_member(member)
     board_id = member.board_list[0].id
     l = get_list_by_list_name(board_id, "To Do")
     add_trello_task(title, target_date, board_id, l.id)
 
-    
+def delete_task(id_long):
+    delete_trello_task(id_long)
+
 
 
 #########tests################################
