+++ todo_app/templates/index.html
@@ -94,7 +94,7 @@
                   {% if item.status=="Done" %}
                     <a href = {{ url_for('update_task', action="reset_task", id_long=item.id_long, last_sort_col=last_sort_col, sort_dir=sort_dir)}}>[Reset Task]</a>
                   {% endif %}
-                  <a href = {{ url_for('delete_task', id_long=item.id_long, last_sorted_col=last_sort_col, sort_dir=sort_dir)}}>[Delete Task]</a>
+                  <a href = {{ url_for('delete_a_task', id_long=item.id_long, last_sorted_col=last_sort_col, sort_dir=sort_dir)}}>[Delete Task]</a>
               </td>
              
             </tr>
